Add a completed Rust port of the solution to problem 96.
This solution drops a lot of the custom sudoku reduction and threading
code the old problem had, but it turns out the new Vec-based exact cover
implementation is fast enough that we don't need any of it. As-is, this
solution is not even close to being one of the top-ten slowest problems.
So, good enough.